move TestSlotWrapper to a __tests__ directory and rename it
