+++ src/lib/components/Link/Link.test.ts
@@ -2,18 +2,18 @@ import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/svelte";
 import { describe, it, expect } from "vitest";
 
-import TestSlotWrapper from "$lib/components/TestSlotWrapper.svelte";
+import SlotWrapper from "$lib/components/__tests__/SlotWrapper.svelte";
 import Link from "./Link.svelte";
 
 const slotContent = "link text";
 
 describe("Link", () => {
   it("renders", () => {
-    render(TestSlotWrapper, { props: { Component: Link, slotContent, href: "/about" } });
+    render(SlotWrapper, { props: { Component: Link, slotContent, href: "/about" } });
     expect(screen.getByText(slotContent)).toHaveAttribute("href", "/about");
   });
   it("renders with an external link", () => {
-    render(TestSlotWrapper, {
+    render(SlotWrapper, {
       props: {
         Component: Link,
         slotContent,
@@ -24,7 +24,7 @@ describe("Link", () => {
     expect(screen.getByText(slotContent)).toHaveAttribute("rel", "external");
   });
   it("renders with an alternate variation", () => {
-    render(TestSlotWrapper, {
+    render(SlotWrapper, {
       props: { Component: Link, slotContent, href: "/", alternate: true },
     });
     // Ideally we would check the color here, but toHaveStyle doesn't work.
